Added multi-series bar graph support.
Added functionality so that multi-series bar graphs are not only capable
of being drawn in the default overlay mode but also drawn in the side by
side mode, which I am calling series_spread. I have added functionality
to both the drawing portion and event handling portion of Flot so that
these side by side multi-series bar graphs can be interacted with in the
same fashion as the normal single series bar graphs. To enable or
disable the side by side mode one simply needs to set the series_spread:
true option in the "bars:" section of the options. I will include an
example demonstrating the appropriate use of this method soon.